Django/history_app_django_backend/requirements.txt
appnope==0.1.0
asgiref==3.2.10
autopep8==1.5.4
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
decorator==4.4.2
Django==3.1.2
django-cors-headers==3.5.0
django-storages==1.10.1
djangorestframework==3.12.1
Flask==1.1.2
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip-autoremove==0.9.1
prompt-toolkit==3.0.7
protobuf==3.13.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
tornado==6.0.4
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
websocket-client==0.48.0
Werkzeug==1.0.1
